Harvey method improvements
- included additional tplot variables for things, also included a first
pass at sane tplot options for things
- First pass at osculating plane calculation.  This looks like it ought
to work, but it, and all the other derivative calculations from k have
issues when |k| is very, very small.  Which leads to:
- Preliminary analysis using this routine shows a serious need for
getting eror analysis automated and cranked into the calculations.
Future work should maybe include flagging, if not suppressing,
derivative calculations (like osc. plane) for k < err_k as calculated
from uncertainties in B and r.  Right now it'scaveat emptor and
preliminary results just aren't trusted unless |k| > about 1e-4 or so
(1/km).  An automated calculation of uncertainties and propogated out to
an uncertainty in the final value could then be used as a gating value
to not report results until k > k_uncertainty.  High on the issue list.